// src/types/common/MessageDataType.ts
export interface MessagesTypes {
  SUCCESSFUL: string;
  INVALID_USER_DETAILS: string;
  INVALID_USER_ORG_DETAILS: string;
  FETCHED_SUCCESS: string;
  USER_ACCOUNT_DELETED: string;
  USER_ACCOUNT_BLOCKED: string;
  USER_ACCOUNT_DISABLED: string;
  USER_NOT_FOUND: string;
  ACCESS_TOKEN_EXPIRED: string;
  JOB_NOT_FOUND: string;
  JOBS_NOT_FOUND: string;
  FAILED_TO_AUTHENTICATE: string;
  UNAUTHORIZED_ACCESS: string;
  JOB_SAVED: string;
  JOB_REMOVED: string;
  JOB_DELETED: string;
  JOB_SUCCESSFULLY_PUBLISHED: string;
  JOB_SUCCESSFULLY_UNPUBLISHED: string;
  JOB_ALREADY_APPLIED: string;
  JOB_SUCCESSFULLY_APPLIED: string;
  JOB_APPLICATION_NOT_FOUND: string;
  JOB_APPLICATION_STATUS_APPLIED: string;
  JOB_APPLICATION_STATUS_SELECTED: string;
  JOB_APPLICATION_STATUS_REJECTED: string;
  JOB_APPLICATION_STATUS_ON_HOLD: string;
  JOB_APPLICATION_STATUS_JOINED: string;
  JOB_APPLICATION_STATUS_SHORTLISTED: string;
  EXTERNAL_JOBS_SAVED_SUCCESSFULLY: string;
  NO_CSV_FILE_FOUND: string;
  INVALID_PAGE_NUMBER: string;
}